Move jpa related classes into separate package
