+++ vaadin-spring/src/main/java/com/vaadin/flow/spring/data/jpa/JpaFilterConverter.java
