+++ vaadin-spring/src/main/java/com/vaadin/flow/spring/data/jpa/package-info.java
@@ -1,0 +1,2 @@
+@org.springframework.lang.NonNullApi
+package com.vaadin.flow.spring.data.jpa;
